# specs/002-todo-phase2/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo API - Phase II
  description: REST API for multi-user todo application
  version: 1.0.0
  contact:
    name: Todo App Support
    email: support@todoapp.local

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todoapp.local
    description: Production server

tags:
  - name: Authentication
    description: User signup, signin, and session management
  - name: Tasks
    description: Task CRUD operations (user-scoped)

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # Request Schemas
    SignupRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: user@example.com
        password:
          type: string
          minLength: 8
          maxLength: 128
          example: securePassword123

    SigninRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          example: securePassword123

    TaskCreateRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 100
          example: Buy groceries
        description:
          type: string
          maxLength: 500
          example: Milk, eggs, bread

    TaskUpdateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          maxLength: 500
        completed:
          type: boolean

    # Response Schemas
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: user@example.com
        name:
          type: string
          nullable: true
          example: John Doe
        created_at:
          type: string
          format: date-time
          example: 2025-12-31T10:00:00Z

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/UserResponse'

    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        title:
          type: string
          example: Buy groceries
        description:
          type: string
          nullable: true
          example: Milk, eggs, bread
        completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: 2025-12-31T10:00:00Z
        updated_at:
          type: string
          format: date-time
          example: 2025-12-31T10:30:00Z

    TaskListResponse:
      type: array
      items:
        $ref: '#/components/schemas/TaskResponse'

    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: User not found

  # Example Values
  examples:
    SignupSuccess:
      value:
        token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          id: 550e8400-e29b-41d4-a716-446655440000
          email: user@example.com
          name: null
          created_at: 2025-12-31T10:00:00Z

    TaskExample:
      value:
        id: 1
        user_id: 550e8400-e29b-41d4-a716-446655440000
        title: Buy groceries
        description: Milk, eggs, bread
        completed: false
        created_at: 2025-12-31T10:00:00Z
        updated_at: 2025-12-31T10:00:00Z

paths:
  ############################################################################
  # AUTHENTICATION ENDPOINTS
  ############################################################################

  /api/auth/signup:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                $ref: '#/components/examples/SignupSuccess'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Invalid email format"
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Email already registered"

  /api/auth/signin:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: Signs in user with email and password, returns JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SigninRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Invalid email or password"

  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Returns the current authenticated user's information
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Not authenticated"

  ############################################################################
  # TASK ENDPOINTS
  ############################################################################

  /api/{user_id}/tasks:
    get:
      tags:
        - Tasks
      summary: List user tasks
      description: Returns all tasks belonging to the specified user
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID (must match authenticated user)
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
              example:
                - $ref: '#/components/examples/TaskExample'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not authorized to access this user's tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Not authorized"

    post:
      tags:
        - Tasks
      summary: Create task
      description: Creates a new task for the specified user
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID (must match authenticated user)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Title must be between 3 and 100 characters"
        '401':
          description: Not authenticated
        '403':
          description: Not authorized

  /api/{user_id}/tasks/{task_id}:
    get:
      tags:
        - Tasks
      summary: Get single task
      description: Returns a specific task by ID
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Task not found"

    put:
      tags:
        - Tasks
      summary: Update task
      description: Updates task title, description, or completion status
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Validation error
        '404':
          description: Task not found

    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Removes a task permanently
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found

  /api/{user_id}/tasks/{task_id}/toggle:
    patch:
      tags:
        - Tasks
      summary: Toggle task completion
      description: Toggles the completed status of a task
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task not found